--- internal/regexparser/state/state_constructors.go
@@ -1,16 +1,0 @@
-package state
-
-func NewStartingState() State {
-	newStartingInspector := StartingInspector{}
-	return State{StateInspector: newStartingInspector}
-}
-
-func NewFinalState() State {
-	newFinalInspector := FinalInspector{}
-	return State{StateInspector: newFinalInspector}
-}
-
-func NewCharState(char rune) State {
-	charInspector := CharInspector{char: char}
-	return State{StateInspector: charInspector}
-}
